svg/interact/scripted/tabindex-focus-flag.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><metadata><h:link href="https://svgwg.org/svg2-draft/interact.html#Focus" rel="help"/></metadata><g id="default-samples"><rect/><svg/><rect focusable="true"/><h:iframe src="resources/blank.htm"/><h:audio controls="controls"/><h:video controls="controls"/><h:canvas tabindex="0"/><circle tabindex="0"/><ellipse tabindex="0"/><image tabindex="0"/><line tabindex="0"/><mesh tabindex="0"/><path tabindex="0"/><polygon tabindex="0"/><rect tabindex="0"/><svg tabindex="0"/><text tabindex="0"><textPath tabindex="0"/></text><unknown tabindex="0"/><use tabindex="0"/><clipPath tabindex="0"/><desc tabindex="0"/><hatch tabindex="0"/><linearGradient tabindex="0"/><meshgradient tabindex="0"/><metadata tabindex="0"/><radialGradient tabindex="0"/><script tabindex="0"/><style tabindex="0"/><title tabindex="0"/></g><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script><![CDATA[
    var root = document.querySelector("svg");

    test(() => {
        root.focus();
        assert_equals(document.activeElement, root);
    }, 'The document root element should be focusable.');

    const defaultList = [
        ['rect', false],
        ['svg', false],
        ['rect[focusable=true]', true],
        ['a', false],
        ['a[href]', true],
        ['a[focusable=false]', false],
        ['iframe', true],
        ['audio[controls]', true],
        ['video[controls]', true],
        ['canvas[tabindex]', true],
        ['circle[tabindex]', true],
        ['ellipse[tabindex]', true],
        ['foreignObject[tabindex]', true],
        ['g[tabindex]', true],
        ['image[tabindex]', true],
        ['line[tabindex]', true],
        ['mesh[tabindex]', true],
        ['path[tabindex]', true],
        ['polygon[tabindex]', true],
        ['rect[tabindex]', true],
        ['svg[tabindex]', true],
        ['switch[tabindex]', true],
        ['text[tabindex]', true],
        ['textPath[tabindex]', true],
        ['tspan[tabindex]', true],
        ['unknown[tabindex]', true],
        ['use[tabindex]', true],
        ['clipPath[tabindex]', false],
        ['defs[tabindex]', false],
        ['desc[tabindex]', false],
        ['hatch[tabindex]', false],
        ['linearGradient[tabindex]', false],
        ['marker[tabindex]', false],
        ['mask[tabindex]', false],
        ['meshgradient[tabindex]', false],
        ['metadata[tabindex]', false],
        ['pattern[tabindex]', false],
        ['radialGradient[tabindex]', false],
        ['script[tabindex]', false],
        ['style[tabindex]', false],
        ['title[tabindex]', false],
        ['symbol[tabindex]', false],
    ];

    for (entry of defaultList) {
        test(() => {
            var element = document.querySelector('#default-samples ' + entry[0]);
            element.focus();
            if (entry[1])
            assert_equals(document.activeElement, element);
            else
            assert_not_equals(document.activeElement, element);
        }, entry[0] + ' should ' + (entry[1] ? '' : 'not ') + 'be focusable.');
    }
    ]]></script></svg>
